VECGEOM-9; First complete, buildable implementation of a trapezoidal shape.
The interface was copied from the existing box, and the implementations were copied from USolids.  At this point there are no vectorized versions for the navigation methods.

Next steps are to implement a simple geometry including a few trapezoidal volumes, and validate the existing (scalar) outputs.  Once this is done, the existing implementations will be vectorized.