[VECGEOM-1] rudimentary random number for cuda compilation
This is implementation is 'just' to get the compilation to succeed
and those properly manage (neither did the previous C++11 only implementation)
the sequence of random number with multiple threads.

Currently this is only used in testing and in GetPointOnSurface, itself
used for testing.  'The right' solution involves changing the interface
of GetPointOnSurface to take a random generator object and/or state object.