Using inline namespaces instead of require the use of a macro.
In user code, this has the effect of support the use
of 'just'
   vecgeom::NameOfClass
in addition (well, instead really) of
   VECGEOM_NAMESPACE::NameOfClass

If the code is compiled by the regular compiler for CPU,
we now have inline namespace named cxx.
If the code is compiled by nvcc (whether it is on the
host or device ... at last for now), we now have the
inline namespace name cuda.

So we have 3 versions of each classes:

   1. vecgeom::cxx::PlacedBox
   2. vecgeom::cuda::PlacedBox (nvcc-host)
   3. vecgeom::cuda::PlacedBox (device)

(The host and device class ought to be binary layout
compatible (so that memcpy equivalent 'work') unless
the class will always be uploaded to the device via
a construct kernel.