[VecGeom-72]; CUDA fixes to Complex class; Torus gives correct results on GPU
cuda 6.0 did not optimize the statements pow(x,1./3) into a cubic root function
and gave NAN when x<0. This is now fixed with the explicit use of the cbrt function.
The TorusBenchmark now gives identical results on CPU + GPU when we compile in Debug mode.
Strangely there is still a problem when we compile in Release mode (cuda 6.0; gcc 4.8.2)
and there possibly is a bug in function csqrtrealargument() on CUDA.