Disable test that breaks build with ROOT disabled
Test shape_testFromROOTFile.cpp, introduced with commit 06e1f150,
remains enabled and fails to compile if -DROOT=OFF is passed to cmake.
The test needs to either be moved together with other ROOT tests, or
changed to not fail to compile when -DROOT=OFF is passed to cmake.
I will leave that choice to the original committer, so I am just
disabling it for now.