[VecGeom-35] New utility: DumpMotherDaughters
This utility takes a .root geometry file and a (mother) volume name as
command line arguments, and returns a dump of the mother volume and
all its daughters.

Ideally it should be kept under userexamples/src, which currently
contains only a CUDA example, and it would require changes to
CMakeLists.txt to get built.

For now it is under test/globalbenchmarks, so it runs under a "make
test", and default values were added in order to pass that test.