Add script to run benchmarks and create figures
The shape_benchmark script runs each shape benchmark a number of times, computes
the average running times, and makes a figure with the results if gnuplot is
available. The script should be run inside the build directory, and it will
annotate each figure with the backend that was used, as well as the name of the
processor where it was run. The arguments are a list of shape names (Box, Orb,
etc), and the number of points and number of jobs per shape can be set in
environment variables. If gnuplot isn't available, only the data files are
created.