[VecGeom-17] Improvements to the benchmarker: multiple performance measurements in a single job (ex: -nstats 25)
New switch added (ex: -nstats 25) controls the number of independent
performance measurements to be taken.  Default is 1, which reproduces
previous behaviour of the benchmarker.

The purpose of such multiple measurements is to study/eliminate the
effect of bad measurement which are due to fluctuations in the state
of the machine used for the benchmark.  For this purpose, the same
input data is to be used in the measurements, for reproducible
results.

Our experience shows that only a small number of random measurements
are affected by the state of the machine, and most measurements tend
to cluster around the expected value.  In most cases, a small number
of affected results does not significantly affect a simple average,
therefore taking a simple average of all multiple measurements is
pretty effective in removing the fluctuations.

For larger Ntracks (longer processing times), the number of affected
results tend to increase.  In such cases, a more careful analysis
should be done instead of a simple average.  It is suggested to look
at the distributions of processing times for a given value of Ntracks,
and check how much the average is larger than the smaller results.