Polycone: fix compile warnings, message typo, and formatting
Change fNz in Polycone to an unsigned integer to avoid signed/unsigned
integer comparison warnings throughout the code. Fix a warning message
that had a typo in it, and change it to lowercase. Remove some of the
dead code in the files and fix formatting of surrounding code to follow
current coding style. Tests were verified to pass on scalar and Vc
backends before merge.