[surface model] Add initial version of BrepCudaManager
Its TransferSurfData will allocate buffers and transfer the contents
of the surface data to the device.

Annotate functions for CUDA

Also replace all std::min found by the compiler with vecCore::math::Min.

Seems to work, but the safety computed by the surface model is *much*
lower for the constructed test geometry...