Fix for extent of Boolean solids.
* Fix for extent of Boolean solids.

This fixes a major bug introduced by the commit 7894c5ee that changed the semantic of VPlacedVolume::Extent to compute the extent of the displaced volume rather than delegate to the unplaced extent calculation. The implementations of UnplacedBooleanVolume<op>::Extent were converting the extents of the Boolean components to the UnplacedBooleanVolume frame, so after the change this conversion was done twice.