Fix for TestVGDMLNavigation and improvements for debugging.
* Fix for TestVGDMLNavigation and improvements for debugging.

This takes over from the error report improvements done in https://gitlab.cern.ch/VecGeom/VecGeom/-/merge_requests/989

Besides better printing, the executable now allows to debug the failed step in both the reference navigator and the tested one. In case of error, the seed of the last shoot direction is printed and can be replayed, so it is not needed to replay the full sequence.