[surface model] Simpler interface to create framed surfaces from vertices
* More correct rounding error and truncation treatment.

testRaytracing is now adapted to compare the final distance reductions with the estimated error given by rounding rather then kTolerance

* Simpler interface to create framed surfaces from vertices

Added the method builder::CreateLocalSurfaceFromVertices taking a vector of maximum 4 vertices and creating the appropriate frame surface, triangular or quadrilateral. The vertices are deduplicated, if the final number is less than 3 the surface is not created and -1 is returned. It is the user responsibility to check the failure when needed.