[VECGEOM-622] Let CMake handle setting of CMP0135
MR !1045 addressed the problem of CMP0135 policy warnings from the
use of ExternalProject for builtin VecCore, but the solution is
only compatible with CMake 3.24 and newer.

Update fix by removing explicit use of DOWNLOAD_TIMESTAMP and only
relying on cmake_minimum_required with a version range to set the
policy appropriately.