[surface model] Fix conical extent
This MR fixes a bug in the calculation of conical extents.

When computing the extent of the right side of a conical surface, the only the surface that is compared against was transformed to the local coordinate system, however, the first side of the right side could already be transformed w.r.t. the left side first surface and therefore also needs to be transformed.

Then, the calculation of the radius of the zphimask was flawed, since the extent is updated on the fly after each surface comparison but used to calculate r0 of the first surface. This is fixed by removing the usage of the extent in the calculation of the radius at 0 and instead transforming the radius at 0 when transforming to the local reference system.

Additionally, the printout of conical and cylindrical surfaces is improved.

* fix cylinder calculation and improve common surface printout

* additional transformation to local system