[surface model] Introduced embedding for framed surfaces
Framed surfaces are embedding if all the frames of daughters produced on the same common surface must be embedded in the frame of the surface.
* Currently the following surfaces are considered possibly non-embedding: all surfaces produced by phi cuts having dphi=kPi; decomposed polyhedra and extruded top/bottom faces; all Boolean surfaces. In future we can reduce this number by performing an actual frame embedding check.

Improved the error detection in testRaytracing (for now only CPU case)
* added a new CrossingSeq struct that stores all crossing distances and states per ray, and can be compared with another. Skipping zero steps is supported.

Consistency fixes and better printout for NavStateTuple

* [formatting] Allow short functions on a single line

* Improved the error detection in testRaytracing (for now only CPU case)

* added a new CrossingSeq struct that stores all crossing distances and states per ray, and can be compared with another. Skipping zero steps is supported.

* Introduced embedding for framed surfaces

Framed surfaces are embedding if all the frames of daughters produced on the same common surface must be embedded in the frame of the surface.

Currently the following surfaces are considered possibly non-embedding: all surfaces produced by phi cuts having dphi=kPi; decomposed polyhedra and extruded top/bottom faces; all Boolean surfaces. In future we can reduce this number by performing an actual frame embedding check.

* Check exiting candidates consistency