[surface model] Fixes for boolean frames and CS. Several fixes for Boolean surfaces on scenes.
In this MR, a change in the exiting of frames is introduced.

For exiting a framed surface, all the parent frames need to be checked whether they are boolean or not. If they are boolean, a thorough check must be performed if the ray is still in the boolean or not because the exited framed surface might be a child of the outer boolean surface but the frames are not overlapping.

This is shown in the example:
![Screenshot_from_2024-04-11_10-31-09](/uploads/15ec4316974b72f73f8ad0a9570e9668/Screenshot_from_2024-04-11_10-31-09.png)

The bottom box surface that is hit and the top box surface are both children of the top boolean surface, which has as a `fDefaultState` the top volume. Thus, before this MR the state when exiting the bottom box surface was set to the `fDefaultState` as top volume but in fact the ray is still within the boolean. Therefore, it must be checked whether the parent frame of the boolean is actually exited and then the exit state must be set either to the boolean (in case the boolean top frame is not exited) or to the top volume (in case the boolean top frame is exited).

* Adding suggestions

* remove deprecated comment

* apply NavIndex change

* fix exit of boolean virtual surfaces, always do a frame check for boolean parents

[geom] Fix candidates for Boolean solids CS that have non-Boolean internal frames.

Common surfaces generated by a Boolean may have daughter frames like the blue frame with the red dot in the picture, where exiting the daughter will not exit on the default parent state, but rather inside the boolean volume itself. In such cases, the CS must be made visible also from the states representing placements of the Boolean volume.
![ugly_bool_boxes](/uploads/2c0b84b659f9d422436622b7261c7f29/ugly_bool_boxes.gif)

* [geom] Fix candidates for Boolean solids CS that have non-Boolean internal frames.

[surface model]  fix navigation state after scene crossing

Follow up to !1107

When a scene is crossed the top_exit_state must be the default state of that new common surface after the scene transition.

* when exiting a scene the default state of the new surface must be used

[surf] Several fixes for Boolean surfaces of scenes.

This contains several fixes for scenes and Boolean surfaces.
* Boolean surfaces can be re-entering if a child frame is not embedded by a parent Boolean frame on the same common surface. Such CS are now added with negative id's in the entering candidate list of the states of those Booleans.
* Re-entering Boolean CS are skipped based on the small distance. This method is not ideal, but I could not find a better approach for now.
* `PushScene` is not called if the state points to a TOP_SCENE
* `DistanceToUnplaced` factored out of the `ComputeStepAndHit`, used for now just in the exiting loop
* Other small corrections