[surface model] First implementation of mixed precision
This MR proposes a first version of running in mixed precision for the surface model.

This is done in the following way:

The solid conversion is now fully done in the compiled `vecgeom::Precision`. In the initialization of the surface data and the masks the data is casted to `Real_t`. Note that in the creation of the surface data there was a storage and an input template parameter but it was not used. Now, the structs have a storage precision `Real_t` and an input precision `Real_i` and the data is casted accordingly at initialization.
The precision of the surface model for the Raytracer can be chosen in `testRaytracing.h`. To allow for the comparison with the fully double-precision solid model, another array of points and directions is created in the testRaytracing both for host and device and then used accordingly.

Running in mixed precision currently is 60% slower than in double precision on GPU (RTX 3080), probably due to huge amount of casts when doing transformations as they are fully in the compiled `vecgeom::Precision`. They will also need to be templated. Currently, for simple geometries, the raytracing works in mixed precision. More complex shapes give traverse errors.
Handling of traverse errors needs to be improved, currently the `Real_t` of the raytracing is used to determine the tolerance. Note that shapes like the torus, Arb4 or even cones might need to be reconsidered if they can be done in single precision (for torus the answer is clearly no). This needs to be improved.

The MR fixes a bunch of errors and warnings when compiling vecgeom fully in single precision, however many warnings still remain and would require further cleaning. Note that running in full single precision is almost 3x faster on the RTX 3080, but gives significant amount of traverse errors.

The tradeoff between doing as many calculations as possible in single precision to improve the speed of GPU, while retaining high accuracy, remains to be investigated.

* set default to double

* fix return type of templated kTolerance

* some cleaning

* first implementation of mixed precision

[surface] Mixed precision transformations

This MR adds a templated version of Transform3D, called Transform3DMP (for mixed precision). In the surface model, only the mixed precision transformation should be used. Using mixed precision, a simple box is ~30% faster on GPU (RTX2070) than full double precision, while having no traverse errors. Note that other shapes, especially booleans, have not been tested yet. Furthermore, it has to be understood why compiling in full single precision is > 2x faster on GPU.

This MR does not change full double precision performance.

The initializer_list had to be introduced since otherwise the direct initialization of a templated transformation was not possible.

* add template parameters

* fix wrong access

* fix clang formatting

* reset default to double

* fix compiler warnings and replace more transformations

* fix places where compiled vecgeom precision was used

* first compiling version

* WIP, not compiling

Fix for rotation check in Transformation3DMP

Frame embedding checking

Touchables may produce framed surfaces on the same common surface with ones produced by their children, but not necessarily embedding them. So when exiting child frames having a parent, navigation wrongly assumes in such case that the parent is also exited, taking the default state of the common surface as the final result.

This MR initiates a detailed frame embedding check. During deduplication, after finding parent frames, the fParent flag is now conditioned on embedding. The non-implemented embedding checks automatically sets embedding to true, to preserve the previous behavior

* Frame embedding unit test

* Added unit test for frame embedding. Embedding now checked for Window-any and most of Ring-Any frames.

* Added frame check prototypes for all frame combinations.

* Added frame embedding checking, currently not specialized and not used

* The entering candidates loop now uses DistanceToUnplaced

[surface model] fix some kTolerance usages for mixed precision

This small MR fixes a few usages of the old kTolerance and introduces the usage of the precision based kTolerance.

Note that there are more compile-time kTolerances used, however, I propose to only change them gradually when investigating failing rays and not all at once, since when fully compiling with single precision, many traverse errors were observed.

* fix some kTolerance usages