[surface model] Mixed precision improvements
This MR does a few improvements to the mixed precision handling:
- it adds the handling of the ConeTolerances in mixed precision. Note that it is not stable to have the ConeTolerance being the same as the kTolerance. We should aim to reduce the kTolerance to lower values in float, to be tested.
- a relative tolerance is added (but not much used yet).
- MakePlusTolerant(0) = tolerance
- a Tolerance() implementation was removed as it was the same as kToleranceDist<>.
- Some signs in the RingMask were switched.

* fix cone precision

* relative tolerance

[surface model] fix incorrectly swapped indices in Transformation3DMP

This MR fixes a wrong layout of the access of the rotation in Transformation3DMP. This was caught due to point locate errors in the latest CMS geometry, which are solved by this MR.

* fix incorrectly swapped indices

[surface model] avoid multiplication which can fail for overflow in CMS

This MR fixes an issue for the `flip_bool` flag in the common surface deduplication. Due to an integer overflow, the flip_bool was incorrectly set to true for CMS. This fixes incorrect propagation.

This also includes some cleaning from leftover printouts for debugging.

* avoid multiplication which can fail for overflow in CMS

[surface model]

This MR fixes the self-entering candidates. Before, it was possible that two non-embedded daughter volumes caused the self_entering flag to be false.

This fixes infinite loops in the latest CMS geometry.

As a reproducer, the following GDML would fail before this fix:[CMS_polyhedron2.gdml](/uploads/b5b5d9a2b3f0e940c70780413d1e29d9/CMS_polyhedron2.gdml)

* set to -isurf

Fix for the transformation used in relocation after exiting a scene.

The transformation used for recomputing the point on the surface when exiting scenes was using a wrong state.

* Fix for the transformation used in relocation after exiting a scene.