[surface model] fix non-embedded overlapping surfaces for exiting rays
This MR fixes the extruding overlaps similar to !1130.

Normally, if a frame is non-embedded, when exiting, the parents are checked thoroughly. However, in case of extruding overlaps, they are non-embedded but the frame of the parent is not hit. If the parent is embedding, it should still be accepted as a valid hit, even if the framed surface is not hit, since the exiting_FS needs to be set to the highest exiting surface.

* fix non-embedded overlapping surfaces for exiting