[surface model] Deep refactoring of ComputeStepAndHit
Both entering and exiting loops now relocate after each frame is found closer.

Added EnterCS method handling finding the closest entering frame and relocating after CS crossing. A frame hit can be provided to bootstrap the search for daughters only.

Child frames are now marked as embedded if this holds for any parent if there are more parent frames on the same side. By default, frames are marked as not embedded. Frames are sorted on the CS by the embedding property, not embedded first.

* Embedding fixes

* Deep refactoring of ComputeStepAndHit

Both entering and exiting loops now relocate after each frame found closer.

Added EnterCS method handling finding the closest entering frame and relocating after CS crossing. A frame hit can be provided to bootstrap the search for daughters only.

Added ExitCS method finding the topmost exited frame and Relocating on the other side after CS crossing. A frame hit can be provided to bootstrap the search among the exiting frames.

Children frames are now marked as as embedded if this holds for any parent, if there are more parent frames on the same side. By default, frames are marked as not embedded. Frames are sorted on the CS by the embedding property, not-embedded first.

* Refactored relocation on common surface. Added DistanceToLocalFS.