[surface model] More fixes for overlap detection
This MR fixes an issue with booleans in the overlap detection. However, the branch cannot be merged before another fix is merged, so this is a draft for now. Plus, it requires cleaning.

The last two commits were needed to fix 'ugly_bool_scene.gdml' and 'CMS_simple_debug.gdml' as they contain booleans that otherwise fail.

Currently, there are very few rays that fail on LHCb (~1 / 15 mio rays). One of them is even stalling:
( `starting point (1604.275183522523, -649.8006919363708, 1360.678850460461) and direction (-0.8485374307382567, 0.5240162347310636, 0.07342488933867887)`
This could be fixed by forbidding entering the same surface with very small steps (to prevent 0 steps)

at the last commit, CMS2026D110 fails at ~130k rays with overlap detection. This is mostly an issue of not having common surfaces in that sense, to be debugged.

* dont exclude for entering daughters. Exclude only if prev_volume = this volume

* fix booleans that might not be exited

* fix overlap handling on GPU

* for relocation in booleans, do push

* more fixes for overlap detection