[surface model] improve error handling for 0 steps
This MR intends to improve the comparison between the solid an the surface model in the testRaytracing. The motivation is to reduce the errors in the cms geometry, since there are surfaces which are closer than 1e-6, so the push of the solid model will cross them, however, they are far enough apart for the common surface algorithm to not consider them as common surfaces.

In this MR, the accept 0 option is now passed as an input parameter `-accept_zeros`.

With this option, 0 steps are not considered as errors. Furthermore, it improves the handling of zero steps, because only after all zero steps, the states should be compared. Due to this, the steps that should be compared should differ between solid and surface model (depending on the 0 steps). This is now taken into account. Now, errors will be printed like this:

![Screenshot_from_2024-06-19_11-18-57](/uploads/76c97b9b9eea6a295b588d8d7889d125/Screenshot_from_2024-06-19_11-18-57.png)

This error is an example of where after a zero step, the result is still wrong, due to an overlap (fixed by the overlap detection, if turned on).

* improve doc

* revert early exit

* improve error handling for 0 steps