[surface model] Fix ring ring embeddedness
Previously, the embedded flag for rings within rings was not always correct, since rings with a phi cut and a rotation just in the XY plane were considered as a general case. By adding a check for just xy rotated ring masks, the flag is set correctly for many cases in CMS.

Note that similar things might apply for the ZPhi Mask.

Side note: this MR also changes an issue in the overlap detection, which should only rely on exiting surface information, not on entering one.
However, this then ruins the printouts for exited common surfaces, so this should be done differently. Therefore, this remains as a draft.

* improve ring ring embedded check

* revert early exit

* improve error handling for 0 steps