[surface model] Fix entering surfaces with two valid solutions
This MR fixes a critical bug that could appear for second order surfaces such as cones or cylinders.

If there are two valid solutions, but must be checked properly. Previously, if the first valid solution failed the frame check, the second valid solution was not tested, even though it could pass the frame check and be the valid solution.

This MR now checks for a second valid solution and if the frame check for the solution with the shorter distance fails, it will also check the other valid solution.

Additionally, for convenience, a `max_cross` option was added to the `testRaytracing.cpp`, to have a simple input parameter which helps in debugging long rays or 0 step rays.

* clean some redundant comments

* fix small logic errors, cleaning, addition of max_cross option

* version with lambda function, not fully functional

* revert early exit

* improve error handling for 0 steps