Removed implementation namespace for GeoManager
Compiler-dependent namespacing raised a particular problem because of the statics attached to this class (gNavIndex, gCompactPlacedVolBuffer), which made these variables unreachable in case GeoManager.h was included in non-device nvcc compilation units. GeoManager is a host class only, but depending on the compiler, it would appear in either cuda or cxx namespace, while these static variables were only defined for cxx.

* Removed implementation namespace for GeoManager

Compiler-dependent namespacing raised a particular problem because of the statics attached to this class (gNavIndex, gCompactPlacedVolBuffer), which made these variables unreachable in case GeoManager.h was included in non-device nvcc compilation units. GeoManager is a host class only, but depending on the compiler, it would appear in either cuda od cxx namespace, while these static variavles were only defined for cxx.