[surface model] fix embedding flag for polyhedrons and polycones
This MR fixes the embedding flag for polyhedrons and polycones.

Since polyhedrons and polycones are similar to boolean unions, the embedding flag must not be set to true by default. 
In fact, the side surfaces are only embedding:
- when the inner surface is convex
- when the outer surface is concave

Thus, we need to check for convexity and concavity and if it is not met, the surface must be marked as non embedding.
This fixes rays in CMS as well as it reduces warnings due to fEmbedding parents with not fEmbedded daughters.

Reproducer small gdmls for the polycones and polyhedron are:
[CMS_simple.gdml](/uploads/f3798b2f42b5ea90f1884035450b678a/CMS_simple.gdml)
[CMS_simple_polyhedron.gdml](/uploads/86a27c2a9c97d294b67bcea728eabed7/CMS_simple_polyhedron.gdml)
which failed for 10 mio rays but are fixed by this MR.

Furthermore, this MR adds asserts in the overlap detection that prevent segfaults and give meaningful err0r messages instead as well as the `max_cross` variable that did not work in the `ValidateCrossing` call.

* add doc

* unify function calls

* fix embedding flag for polyhedrons and polycones, add assert to prevent segfault in overlap detection, fix max_cross in validatecrossing function