[surface model] pass highest exiting and entering information + fix overlap detection for 0 steps
For the relocation in the overlap detection, only the highest exiting frame information is needed. However, to mark the correct surfaces as overlapping, the final entering surface information is needed.

This MR proposes to pass both (they are stored in a new struct containing both `FSlocator`s), so both are available and can be used in the `testRaytracing.cpp`.

Additionally, the overlap detection was struggling in case of 0 steps. Now, if there are 0 steps, a push is used, fixing the issue.
Now, CMS does not crash with the overlap detection for 1 mio rays and reports ~ half of the errors it does without overlap detection.

* more appropriate naming

* unify exit_FS and hit_FS

* add push for 0 steps in overlap detection

* pass highest exiting and entering information