[surface model] Fix reentering booleans
This MR fixes several issues, mostly related to re-entering booleans.
Previously, in case of boolean unions, the fEmbedded daughters (green box) are not entered through the common surface.

The problem is that the parent frame is a virtual one, so it is not entered and therefore, since the daughter was embedded, the daughter was not entered either.
This is now handled by having a virtual parent flag. In that case, the state is added to the candidates even if it is embedded and then later it is checked, even if no parent is found yet.

Additionally, the ZPhiMask frame check was fixed, there was a mixup with frame1 and frame2 and the transformations are needed in case of rotations in the XY plane (similar to the ring-ring comparison).
Also, the `max_cross` option was fixed, which got lost in merge conflicts after the BVH merge.

![image](/uploads/fa30833b1a00ca28bcf1e88489aa5b30/image.png)

* remove redundant check

* fix max_cross

* fix ZPhi frame mask comparison and clang format

* remove of const was correct

* fix missing const

* fix reentering by introducing virtualparents

* work in progress