[surface model] Restructure two solutions for cones
Previously, it was assumed that if two solutions were found for cones etc that the surfhit was true for both sides.
However, this assumption is not always true, therefore we need to store the result and act accordingly.

The code was also cleaned and clearer comments explain the logic now.

* restructure two solutions