[surface model] Fix booleans with are very close to the next surface
This fixes an error where the overlap detection was triggered incorrectly.

In the case of exiting a boolean, which is less than the push distance to the next boolean, the push would end up outside of the boolean, leading to a detected overlap. In this case, we need to do an `or` of the inside call with and without the push. This must not be done in the case when we check whether we exit the same boolean because then this inside check without the push would always return true although in fact we do exit the boolean.

This fixes this ray in in this gdml [CMS_simple13.gdml](/uploads/e18223ef0fd6cdb308e306d46a98b8c2/CMS_simple13.gdml)
```
./testRaytracing -gdml_name ../../gdml/CMS_simple13.gdml -nrays 7000000 -ongpu 0 -detect_overlaps 1 -mmunit 0.1 -verbosity 0 -debug 1  -accept_zeros 1 -direction 0.9542659949562331, -0.1483842204337957, -0.2595352268892319 -max_cross 5 -point -7619.809842534158, 1058.923587696931, 1249.52834086188
```

* fix self entering plus comments

* exclude selfentering booleans from inside with or without push

* problem if 0 step comes next..