[surface model] Fix CS generation for top scene surfaces and using the correct point for self-entering booleans
This MR fixes two issues:

In the common surface creation, if both are top scene surfaces, they should be put on the same CS and not do the check by states on the same side (because maybe the other side doesn't exist, but still those surfaces need to be on the same common surface).

This fixes the following ray in this geometry [CMS_simple15.gdml](/uploads/0c5733e343a307ada4bff98cdf426485/CMS_simple15.gdml):

```
./testRaytracing -gdml_name ../../gdml/CMS_simple15.gdml -ongpu 0 -detect_overlaps 1 -mmunit 0.1 -verbosity 1 -debug 1  -accept_zeros 1   -min_per_scene 1 -point 110.89155855880512  65.00385387256534   2.49695477886371 -direction -0.6411200753352351, -0.1373176391492646, -0.7550555707897365
```

Furthermore, for self-entering booleans in scenes, not the scene point needs to be used but the point in the volume. This should be better understood, (especially because something similar in the BVH might be missing), but for now this is giving the correct results in CMS.

This fixes the following ray of the same geometry as above:
```
./testRaytracing -gdml_name ../../gdml/CMS_simple15.gdml -ongpu 0 -detect_overlaps 1 -mmunit 0.1 -verbosity 1 -debug 1  -accept_zeros 1   -min_per_scene 1 -point -8808.413339786459, 8192.275015981819, -4141.180381555163 -direction 0.7034932057269896, -0.6328155030451438, 0.3234839232507317
```

Last, the printout of the hit surface is improved for debugging, now also printing the frame id and the side.

* fix CS generation for top scene surfaces and using the correct point for self-entering booleans in scenes