[surface model] Remove noframe boxes from BVH
Added another array in volume shells, containing indices of real framed surfaces in the fSurfacesArray. This array is passed to the BVH constructions and then used in navigation.

* Remove noframe boxes from BVH

Added another array in volume shells, containing indices of real framed surfaces in the fSurfacesArray. This array is passed to the BVH constructions and then used in navigation.