[surface model] Mark step as 0 step if close to other 0 step
Previously, a 0 step was marked as a 0 step if it was smaller than 1e-5. However, due to small differences, the solid model could end up with 1.005e-5 and the surface model with 0.995e-5. Although both steps are only 1e-7 from each other, one was marked as 0 step and the other one was not, leading to errors in the evaluation.

This is fixed by checking if one step is a 0 step, if the other step is less than 1e-6 from the 0 step and if it is, it is also marked as 0 step.

This MR fixes the following ray with this geometry: [CMS_simple18.gdml](/uploads/38c527b12d66c8b0091f3d9ef3b2d802/CMS_simple18.gdml)

```
./testRaytracing -gdml_name ../../gdml/CMS_simple18.gdml -nrays 10000000 -ongpu 0 -detect_overlaps 1 -mmunit 0.1 -verbosity 0 -debug 1  -accept_zeros 1 -direction -0.3984079029182088, 0.5041991222135486, 0.7661947455127855 -point -57.48186150143374 145.62316004284003 -36.07798265824022  -max_cross 4
```

* mark step as 0 step if close to other 0 step