[surface model] Parent can be virtual for any boolean
Previously, I assumed that virtual parents only occur in case if the parent is boolean and the frame exists on both sides. However, it doesn't need to exist on both sides, as this example from CMS shows.
The daughter shares a common surface with the boolean parent. For the boolean, this is a virtual frame, while for the daughter, it is a real one. Therefore, the state of the daughter needs to be added and checked even if it is embedded.

This fixes the following ray in this simple reproducer from CMS: [CMS_simple20.gdml](/uploads/3d37c6f4de6e2da78e028c90beed7a28/CMS_simple20.gdml)

```
./testRaytracing -gdml_name ../../gdml/CMS_simple20.gdml -nrays 1000000 -ongpu 0 -detect_overlaps 1 -mmunit 0.1 -verbosity 1 -debug 1  -accept_zeros 1 -point -436.2093795020209, -154.5969709494438, -345.8886762700702 -direction 0.4237314512047304, 0.1855310071220027, 0.8865832745186594
```

This fixes rays in CMS and I checked, the runtime for 15 mio rays does not change measurably with this change.

* fix unused variable

* parent can be virtual for any boolean