[surface model] Calculate safety in distance calculation and use this for rejection of negative distances
Previously, a ray was rejected if the distance was smaller than `-vecgeom::kTolerance`. However, in the inside check, we use the safety to calculate if a point is inside the solid. This could lead to the case where the safety found the point to be inside (with < -1e-9 safety distance), but the distance was rejected because it exceeded the threshold ( = -1.5e-9), which caused wrong behaviour.

In fact, we should use the safety to do this check for a too small distance.

This MR adds the safety calculation (only the absolute is needed). In the current implementation, we still do negative steps (we could just set it to 0 instead of pushing backwards). The check whether the safety needs to be calculated (only if the distance is negative and only if it is not so negative that it cannot be within the safety) could be improved but in the current version, no run time difference could be observed (so it is calculated only very rarely).

This kind of check could also be added to the torus, which currently still uses a distance to reject hits (but it needs to be quite negative to be rejected, so very unlikely at the moment).

This MR fixes then the following ray in the following geometry: [CMS_simple17.gdml](/uploads/35cca1a01bd471d422f9399f47a4f7f1/CMS_simple17.gdml)

```
./testRaytracing -gdml_name ../../gdml/CMS_simple17.gdml -nrays 1000000 -ongpu 0 -detect_overlaps 1 -mmunit 0.1 -verbosity 0 -debug 1  -accept_zeros 1 -point  118.58717365995562 -73.35318234807073 121.7676173435666  -direction -0.3427042096998675, -0.481832929359343, -0.8064681350425346 -max_cross 3
```

* calculate safety in distance calculation and use this for rejection of negative distances