[surface model] Fix scenes with BVH
This MR fixes a few issues with the BVH and scenes:

Sometimes, the state is still | /TOP_SCENE at the end of the navigation, this needs to be popped.
Furthermore, similarly to the loop navigator, if `is_scene` we need to use a different point. Note that in the normal navigator there is also an exclusion for self-entering booleans, so maybe this is still missing.

At this state, CMS runs 1 mio rays without errors, but runs into some other error for 10 mio rays (this will be fixed in a separate MR).

* fix scenes with BVH